backend/requirements.txt
Django>=5.0,<5.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
